refactor: single Rust CLI implementation shared by cargo and pip
Move CLI logic from bin/xa11y.rs into a #[doc(hidden)] cli module in
the xa11y crate. The binary becomes a thin wrapper calling cli::run().

PyO3 exports _cli_main() which calls the same Rust code. The Python
_cli.py is now a 3-line trampoline instead of a full reimplementation.

This eliminates the duplicated CLI logic between Rust and Python —
both `cargo install xa11y` and `pip install xa11y` run identical code.
